requirements.txt
alembic==1.10.3
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
filelock==3.11.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pipenv==2023.3.20
platformdirs==3.2.0
pycodestyle==2.10.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
WTForms==3.0.1
